Arbitrary Dimensional Clifford Algebras

A suite of routines for Clifford algebras, using the 'Map' class of the Standard Template Library. Canonical reference: Hestenes (1987, ISBN 90-277-1673-0, "Clifford algebra to geometric calculus"). Special cases including Lorentz transforms, quaternion multiplication, and Grassmann algebra, are discussed. Vignettes presenting conformal geometric algebra, quaternions and split quaternions, dual numbers, and Lorentz transforms are included. The package follows 'disordR' discipline.

Details

The DESCRIPTION file: This package was not yet installed at build time.
Index: This package was not yet installed at build time.

Author

Robin K. S. Hankin [aut, cre] (<https://orcid.org/0000-0001-5982-0415>)

Maintainer: Robin K. S. Hankin <hankin.robin@gmail.com>

References

J. Snygg (2012). A new approach to differential geometry using Clifford's geometric Algebra, Birkhauser. ISBN 978-0-8176-8282-8

D. Hestenes (1987). Clifford algebra to geometric calculus, Kluwer. ISBN 90-277-1673-0

C. Perwass (2009). Geometric algebra with applications in engineering, Springer. ISBN 978-3-540-89068-3

D. Hildenbrand (2013). Foundations of geometric algebra computing. Springer, ISBN 978-3-642-31794-1

See also

Examples

as.1vector(1:4)
#> Element of a Clifford algebra, equal to
#> + 1e_1 + 2e_2 + 3e_3 + 4e_4

as.1vector(1:4) * rcliff()
#> Element of a Clifford algebra, equal to
#> + 5e_1 + 10e_2 + 27e_12 + 15e_3 - 18e_13 + 9e_23 + 20e_4 - 36e_1234 + 15e_5 -
#> 2e_15 - 4e_25 + 18e_35 + 5e_135 + 10e_235 - 26e_45 - 20e_345 - 6e_1345 -
#> 12e_2345 + 1e_16 + 2e_26 + 31e_36 - 17e_46 - 7e_1346 - 14e_2346 + 4e_256 -
#> 2e_456 - 1e_12456 + 3e_23456

# Following from Ablamowicz and Fauser (see vignette):
x <- clifford(list(1:3,c(1,5,7,8,10)),c(4,-10)) + 2
y <- clifford(list(c(1,2,3,7),c(1,5,6,8),c(1,4,6,7)),c(4,1,-3)) - 1
x*y # signature irrelevant
#> Element of a Clifford algebra, equal to
#> - 2 - 4e_123 - 16e_7 + 8e_1237 - 6e_1467 - 12e_23467 + 2e_1568 + 4e_23568 +
#> 10e_6710 - 40e_235810 - 30e_456810 + 10e_157810