Class “dot”

The dot object is defined so that idiom like .[x,y] returns the commutator, that is, (x*y-y*x)/2. The factor of 2 ensures that .[x,y] == x %X% y.

The dot object is generated by running script inst/dot.Rmd, which includes some further discussion and technical documentation, and creates file dot.rda which resides in the data/ directory.

Arguments

x

Object of any class

i,j

elements to commute

...

Further arguments to dot_error(), currently ignored

Value

Always returns an object of the same class as xy.

Author

Robin K. S. Hankin

Examples

x <- rcliff()
y <- rcliff()
z <- rcliff()

.[x,.[y,z]] + .[y,.[z,x]] + .[z,.[x,y]]  # Jacobi identity
#> Element of a Clifford algebra, equal to
#> the zero clifford element (0)